Source code/requirements.txt
absl-py==2.1.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
aspose-words==24.10.0
Aspose.Slides==24.10.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
attrs==24.2.0
audioread==3.0.1
babel==2.16.0
bleach==6.2.0
blinker==1.8.2
blis==1.1.0
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.1
click==8.1.7
cloudpathlib==0.20.0
colorama==0.4.6
comm==0.2.2
commonmark==0.9.1
confection==0.1.5
contourpy==1.3.0
cryptography==44.0.0
cvzone==1.6.1
cycler==0.12.1
cymem==2.0.10
debugpy==1.8.9
decorator==5.1.1
defusedxml==0.7.1
diffusers==0.31.0
dill==0.3.9
docx2txt==0.8
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.21.1
ffmpeg-python==0.2.0
filelock==3.16.1
filterpy==1.4.5
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
flatbuffers==24.3.25
fonttools==4.54.1
fqdn==1.5.1
fsspec==2024.10.0
future==1.0.0
gast==0.6.0
gcloud==0.18.3
geocoder==1.38.1
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
googleapis-common-protos==1.66.0
googletrans==3.0.0
greenlet==3.1.1
grpcio==1.68.1
gTTS==2.5.3
h11==0.14.0
h2==3.2.0
h5py==3.12.1
hpack==3.0.0
hstspreload==2024.10.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.26.1
hyperframe==5.2.0
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.28.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jax==0.4.35
jaxlib==0.4.35
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-core==5.7.2
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.3.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
jwcrypto==1.5.6
kagglehub==0.3.6
keras==3.7.0
kiwisolver==1.4.7
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
llvmlite==0.43.0
lxml==5.3.0
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.14
mistune==3.0.2
ml-dtypes==0.4.1
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.0
murmurhash==1.0.11
mysql-connector==2.2.9
namex==0.0.8
narwhals==1.20.1
nbclient==0.10.1
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
norfair==2.2.0
notebook==7.3.1
notebook_shim==0.2.4
numba==0.60.0
numpy==2.2.1
oauth2client==4.1.3
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pdf2image==1.17.0
pdfminer.six==20240706
pdfminer3==2018.12.3.0
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1733327343728/work
pillow==11.0.0
platformdirs==4.3.6
plotly==5.24.1
pooch==1.8.2
preshed==3.0.9
proglog==0.1.10
prometheus_client==0.21.1
prompt_toolkit==3.0.48
protobuf==5.29.2
psutil==6.1.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.4
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.18.0
PyMySQL==1.1.1
pyparsing==3.2.0
Pyrebase4==4.8.0
pyresparser==1.0.6
pyrsistent==0.20.0
python-dateutil==2.9.0.post0
python-json-logger==3.2.1
python-jwt==4.1.0
python-pptx==1.0.2
python-version==0.0.2
pytz==2024.2
pywin32==308
pywinpty==2.0.14
PyYAML==6.0.2
pyzmq==26.2.0
ratelim==0.1.6
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==0.10.1
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
safetensors==0.4.5
scikit-learn==1.3.0
scipy==1.14.1
seaborn==0.13.2
Send2Trash==1.8.3
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.1
soundfile==0.12.1
soupsieve==2.6
soxr==0.5.0.post1
spacy==3.8.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.11.0
SQLAlchemy==2.0.36
srsly==2.5.0
stack-data==0.6.3
streamlit==1.41.1
streamlit-tags==1.2.8
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
terminado==0.18.1
thinc==8.3.3
threadpoolctl==3.5.0
tinycss2==1.4.0
tkintertable==1.3.3
tokenizers==0.20.3
toml==0.10.2
tomli==2.2.1
torch==2.5.1
torchaudio==2.4.0
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.46.3
typer==0.15.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.48
ultralytics-thop==2.0.13
uri-template==1.3.0
urllib3==1.26.20
Wand==0.6.13
wasabi==1.1.3
watchdog==6.0.0
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.13
wrapt==1.17.0
WTForms==3.2.1
XlsxWriter==3.2.0
zipp==3.21.0
